Move MoyaError comparison function.
Relocated `MoyaError.sameErrorType()` to an extension in the test target.
